chat_app/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
anthropic